Remove duplicate and outdated test files
Deleted duplicate and outdated test files from dist and src/tests directories. Added new test files to tests directory and updated references to improve test organization.

+++ src/controllers/authController.js
@@ -220,18 +220,20 @@ const confirmUser = async (req, res, next) => {
 
 
 //user a lib such as node-cron for cleaning unconfirmed users after a setted time
-
-cron.schedule('0 * * * *', async () => {
-    const dateLimit = new Date(Date.now() - (1000 * 60 * 60))
-    try {
-        const result = await temporaryUser.deleteMany({
-            isEmailConfirmed: false,
-            createdAt: {$lt: dateLimit}
-        })
-        console.log('Deleted users', result.deletedCount)
-    } catch (error) {
-        console.log('An error ocurred while trying to delete temporary users')
-    }
-})
+if (process.env.NODE_PRODUCTION === 'true') {
+    cron.schedule('0 * * * *', async () => {
+        const dateLimit = new Date(Date.now() - (1000 * 60 * 60))
+        try {
+            const result = await temporaryUser.deleteMany({
+                isEmailConfirmed: false,
+                createdAt: {$lt: dateLimit}
+            })
+            console.log('Deleted users', result.deletedCount)
+        } catch (error) {
+            console.log('An error ocurred while trying to delete temporary users')
+        }
+    })
+}
 
 module.exports = {register, logIn, changePassword, logout, confirmUser, generateTemporaryUser}
+
